+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.19.3
 - State: Create form hide id if noEdit 
+- Adapter API: query parameter
 
 # 0.19.2
 - Action conditions